worker/requirements.txt
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0